Refactor rename TaskTracks => TaskTrack 2
